+++ server/app.js
@@ -1,21 +1,33 @@
-const express = require('express');
-const graphqlHTTP = require('express-graphql').graphqlHTTP;
+const pwd = require("./local");
+const express = require("express");
+const graphqlHTTP = require("express-graphql").graphqlHTTP;
+const mongoose = require("mongoose");
+
+const schema = require("./schema/schema");
+const testSchema = require("./schema/types_schema");
+
+const app = express();
 
 /*
 mongodb+srv://fr:<password>@cluster0.qfcnchr.mongodb.net/test
 */
 
-const schema = require('./schema/schema')
-const testSchema = require('./schema/types_schema')
+mongoose.connect(
+  "mongodb://fr:" + pwd.pwd + ">@cluster0.qfcnchr.mongodb.net/test"
+);
+mongoose.connection.once("open", () => {
+  console.log("Yes! We are connected!");
+});
 
-const app = express();
-
-app.use('/graphql', graphqlHTTP({
+app.use(
+  "/graphql",
+  graphqlHTTP({
     graphiql: true,
-    schema: testSchema
-}));
+    schema: testSchema,
+  })
+);
 
-
-app.listen(4000, () =>{ //localhost:4000
-    console.log('Listening for requests on my awesomne port 4000');
-})
+app.listen(4000, () => {
+  //localhost:4000
+  console.log("Listening for requests on my awesomne port 4000");
+});